feat(US-0005108): Complete first on premise deploy version
- Added tests
- Refactor
- Removed un-used test classes
